requirements.txt
asgiref==3.7.2
Django==4.2.3
factory-boy==3.3.0
Faker==19.3.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
